requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
boto3
pypdf2
langchain
langchain-community
langchain-openai
langchain-huggingface
sentence-transformers
transformers
torch
accelerate
bitsandbytes
faiss-cpu
python-dotenv
streamlit
mlflow
nemoguardrails
ragas
presidio-analyzer
presidio-anonymizer
pytest
httpx
langgraph
